\name{CompetingRiskFrailty-internal}
\alias{CompetingRiskFrailtyOptim}
\title{Internal function for fitting of the comepting-risks-with-frailties survival model}
\description{The function 'CompetingRiskFrailtyOptim' implements an optimization procedure and 
is used internally in the body of the 'CompetingRiskFrailtySurvfitCreate' function.}
\details{This function is not to be called by the user.}
\keyword{internal}

